fix: do not set previousActiveElement for toasts
