+++ src/constants.js
@@ -1,0 +1,1 @@
+export const RESTORE_FOCUS_TIMEOUT = 100
