+++ src/globalState.js
@@ -1,3 +1,5 @@
+import { RESTORE_FOCUS_TIMEOUT } from './constants'
+
 const globalState = {}
 
 export default globalState
@@ -11,7 +13,7 @@ export const restoreActiveElement = () => {
       globalState.previousActiveElement.focus()
       globalState.previousActiveElement = null
     }
-  }, 100) // issues/900
+  }, RESTORE_FOCUS_TIMEOUT) // issues/900
   if (typeof x !== 'undefined' && typeof y !== 'undefined') { // IE doesn't have scrollX/scrollY support
     window.scrollTo(x, y)
   }
